feature : create common package in api module (advice&dto) #6
