+++ Trim-Api/src/main/java/trim/api/domains/comment/vo/request/CreateQuestionCommentRequest.java
@@ -1,0 +1,19 @@
+package trim.api.domains.comment.vo.request;
+
+import lombok.Getter;
+import trim.domains.comment.domain.QuestionComment;
+import trim.domains.member.domain.Member;
+import trim.domains.question.domain.Question;
+
+@Getter
+public class CreateQuestionCommentRequest {
+    private String content;
+
+    public QuestionComment from(Member writer, Question question){
+        return QuestionComment.builder()
+                .content(this.getContent())
+                .question(question)
+                .writer(writer)
+                .build();
+    }
+}
