+++ Trim-Api/src/main/java/trim/api/domains/member/controller/MemberApiController.java
@@ -1,0 +1,19 @@
+package trim.api.domains.member.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.*;
+import trim.api.domains.member.service.RegisterMemberUseCase;
+import trim.api.domains.member.vo.MemberRequest;
+
+@RequiredArgsConstructor
+@RestController
+@RequestMapping("/api/members")
+public class MemberApiController {
+
+    private final RegisterMemberUseCase registerMemberUseCase;
+
+    @PostMapping
+    public Long registerServerTestMember(@RequestBody MemberRequest memberRequest) {
+        return registerMemberUseCase.execute(memberRequest);
+    }
+}
