+++ Trim-Api/src/main/java/trim/api/domains/member/vo/MemberRequest.java
@@ -1,0 +1,18 @@
+package trim.api.domains.member.vo;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import trim.domains.member.domain.Role;
+import trim.domains.member.domain.SocialType;
+
+@Getter
+@RequiredArgsConstructor
+public class MemberRequest {
+
+    private final String username;
+    private final String nickname;
+    private final String email;
+    private final String socialType;
+    private final String role;
+
+}
