--- Trim-Api/src/main/java/trim/api/question/dto/response/FindQuestionResponse.java
@@ -1,19 +1,0 @@
-package trim.api.question.dto.response;
-
-import lombok.Builder;
-import lombok.Getter;
-import trim.domains.question.domain.Question;
-
-@Getter
-@Builder
-public class FindQuestionResponse {
-    private String title;
-    private String nickname;
-
-    public static FindQuestionResponse of(Question question){
-        return FindQuestionResponse.builder()
-                .title(question.getTitle())
-                .nickname(question.getWriter().getNickname())
-                .build();
-    }
-}
