+++ Trim-Api/src/test/java/trim/api/question/QuestionUseCaseTest.java
@@ -5,17 +5,16 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import trim.TrimApiServiceApplication;
-import trim.api.question.dto.request.QuestionRequest;
-import trim.api.question.dto.response.FindQuestionResponse;
-import trim.api.question.dto.response.QuestionResponse;
-import trim.api.question.service.*;
+import trim.api.domains.question.dto.request.QuestionRequest;
+import trim.api.domains.question.dto.response.FindQuestionResponse;
+import trim.api.domains.question.dto.response.QuestionResponse;
+import trim.api.domains.question.service.CreateQuestionUseCase;
+import trim.api.domains.question.service.FindAllQuestionUseCase;
+import trim.api.domains.question.service.FindQuestionUseCase;
 import trim.domains.comment.adaptor.QuestionCommentAdaptor;
-import trim.domains.comment.domain.QuestionComment;
 import trim.domains.member.domain.Member;
 import trim.domains.question.adaptor.QuestionAdaptor;
 import trim.domains.question.domain.Question;
